# BackendGeoprofs/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Geoprofs",
        "description": "API documentatie voor geoprofs verlofsregistratiesysteem",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local api server"
        }
    ],
    "paths": {
        "/auth/request": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get the currently logged-in user",
                "operationId": "bed316293511b21d73531c2926b9ce36",
                "responses": {
                    "200": {
                        "description": "Authenticated user returned"
                    },
                    "401": {
                        "description": "No logged-in user"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login with username and password",
                "operationId": "25ae12e27b55dd95a5288e572e51f84e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "testuser"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged-in user returned"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout and invalidate session",
                "operationId": "9cc6452d04fb264f59efe7ded58d98de",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "No logged-in user"
                    }
                }
            }
        },
        "/user/{userId}/verlofaanvraag": {
            "get": {
                "tags": [
                    "Verlofaanvraag"
                ],
                "summary": "Get all verlofaanvragen of a specific user",
                "operationId": "6f05f5929cf2d0a4571b3cd902fda69b",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of verlofaanvragen retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Verlofaanvraag"
                ],
                "summary": "Create a new verlofaanvraag for a user",
                "operationId": "fe8a31fde554bd7aac25b216c5e39def",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Verlofaanvraag details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reden",
                                    "startdatum",
                                    "einddatum",
                                    "status"
                                ],
                                "properties": {
                                    "reden": {
                                        "type": "string",
                                        "example": "Vakantie"
                                    },
                                    "startdatum": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    },
                                    "einddatum": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-10"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "In behandeling"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Verlofaanvraag successfully created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/user/{userId}/verlofaanvraag/{id}": {
            "get": {
                "tags": [
                    "Verlofaanvraag"
                ],
                "summary": "Get a specific verlofaanvraag of a user",
                "operationId": "175154480d8f3d8dfe15c2105e7befd3",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the verlofaanvraag",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verlofaanvraag retrieved successfully"
                    },
                    "404": {
                        "description": "Verlofaanvraag not found"
                    }
                }
            }
        },
        "/user/{userId}/verlofaanvraag/{id}/approve": {
            "put": {
                "tags": [
                    "Verlofaanvraag"
                ],
                "summary": "Approve a verlofaanvraag (admin only)",
                "operationId": "2aa3321cf1ebedec7cacc30102bd8c06",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "description": "Approval status update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "Goedgekeurd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verlofaanvraag approved successfully"
                    },
                    "403": {
                        "description": "Forbidden (not an admin)"
                    }
                }
            }
        },
        "/user/{userId}/verlofaanvraag/{id}/reject": {
            "put": {
                "tags": [
                    "Verlofaanvraag"
                ],
                "summary": "Reject a verlofaanvraag (admin only)",
                "operationId": "86566af1400887ed438c5cce28b68bfa",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "description": "Rejection reason and status update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status",
                                    "afkeuringsreden"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "Afgekeurd"
                                    },
                                    "afkeuringsreden": {
                                        "type": "string",
                                        "example": "Te weinig vakantiedagen"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verlofaanvraag rejected successfully"
                    },
                    "403": {
                        "description": "Forbidden (not an admin)"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Verlofaanvraag",
            "description": "Verlofaanvraag"
        }
    ]
}